fix: replace sqlcipher dependencies to point to https://github.com/OutSystems/android-database-sqlcipher.git#feat/RNMT-2706_64bit
Add 64bit architecture files

Opening old databases (created with version 3.4.0) with the new native "key" methods fails, most probably related to changes related to native_key methods in SQLiteDatabase.cpp file.
To fix this issue, when first failing to open the database, the system will try to open the database using the old key method. If it succeeds to open the DB then it will change the password to the same password but now using the new key methods. This way, this error never occurs again to that instance of the DB.